Add option to disable the IO cache reset for every new frame.
This is necessary if architectures need the context they
are embedded in.
Context means: Instruction word +- x bytes.
For Hexagon the disassembly and IL steps are incorrect,
if the context is not the same every time and differs
from the actual binary.
Because instructions disassemble differently depending on
their neighbors.
If the neighbor instruction of a packet indicated the
start of it, and during the IL step execution this
neighboring instructions is no longer the same,
the interpretation is different and the IL step is invalid.