Add a post-op list with all registers and their count.
If the post-operand list had a register twice in it and
the last entry had the correct value, a mismatch
happened anyway.
Because the first value was tested and did not match.

This is not correct, since multiple writes to a register can happen.
Only the last post-operand should be considered.